Make app it's own widget that contains scoreboardhome widget as its content
